MAINT: fix usage of aliases to builtin types
NumPy v1.20.0 deprecated the usage aliases like np.int, np.float,
np.complex, np.str, etcetera.

See: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations

+++ lmfit/minimizer.py
@@ -1942,10 +1942,10 @@ class Minimizer:
 
             if len(result.var_names) == 1:
                 grid_result = np.array([res for res in zip(zip(grid_points), grid_score)],
-                                       dtype=[('par', 'O'), ('score', 'float64')])
+                                       dtype=[('par', 'O'), ('score', 'float')])
             else:
                 grid_result = np.array([res for res in zip(zip(*grid_points), grid_score)],
-                                       dtype=[('par', 'O'), ('score', 'float64')])
+                                       dtype=[('par', 'O'), ('score', 'float')])
             grid_result_sorted = grid_result[grid_result.argsort(order='score')]
 
             result.candidates = []
